+++ src-docs/src/views/html_id_generator/html_id_generator.js
@@ -1,4 +1,4 @@
-import React, { Component, Fragment } from 'react';
+import React, { useState, Fragment } from 'react';
 
 import {
   EuiFlexGroup,
@@ -9,35 +9,26 @@ import {
 
 import { htmlIdGenerator } from '../../../../src/services';
 
-export default class extends Component {
-  constructor(props) {
-    super(props);
+export default () => {
+  const [value, setValue] = useState(htmlIdGenerator()());
 
-    this.state = {
-      value: htmlIdGenerator()(),
-    };
-  }
-
-  reGenerate = () => {
-    this.setState({ value: htmlIdGenerator()() });
+  const reGenerate = () => {
+    setValue(htmlIdGenerator()());
   };
 
-  render() {
-    const { value } = this.state;
-    return (
-      <Fragment>
-        <EuiFlexGroup
-          justifyContent="flexStart"
-          gutterSize="m"
-          alignItems="center">
-          <EuiFlexItem grow={false}>
-            <EuiCode>{value}</EuiCode>
-          </EuiFlexItem>
-          <EuiFlexItem grow={false}>
-            <EuiButton onClick={this.reGenerate}>Regenerate</EuiButton>
-          </EuiFlexItem>
-        </EuiFlexGroup>
-      </Fragment>
-    );
-  }
-}
+  return (
+    <Fragment>
+      <EuiFlexGroup
+        justifyContent="flexStart"
+        gutterSize="m"
+        alignItems="center">
+        <EuiFlexItem grow={false}>
+          <EuiCode>{value}</EuiCode>
+        </EuiFlexItem>
+        <EuiFlexItem grow={false}>
+          <EuiButton onClick={reGenerate}>Regenerate</EuiButton>
+        </EuiFlexItem>
+      </EuiFlexGroup>
+    </Fragment>
+  );
+};