+++ src-docs/src/views/html_id_generator/html_id_generator_prefix.js
@@ -1,4 +1,4 @@
-import React, { Component, Fragment } from 'react';
+import React, { useState, Fragment } from 'react';
 
 import {
   EuiFieldText,
@@ -8,47 +8,37 @@ import {
   EuiCode,
   EuiFormRow,
 } from '../../../../src/components';
+
 import { htmlIdGenerator } from '../../../../src/services';
 
-export class HtmlIdGeneratorPrefix extends Component {
-  constructor(props) {
-    super(props);
+export const HtmlIdGeneratorPrefix = () => {
+  const [prefix, setPrefix] = useState('Id');
+  const [customId, setCustomId] = useState(htmlIdGenerator('Id')());
 
-    this.state = {
-      prefix: 'Id',
-      id1: htmlIdGenerator('Id')(),
-    };
-  }
-
-  onSearchChange = e => {
+  const onSearchChange = e => {
     const prefix = e.target.value;
-    this.setState({
-      prefix,
-      id1: htmlIdGenerator(prefix)(),
-    });
+    setPrefix(prefix);
+    setCustomId(htmlIdGenerator(prefix)());
   };
 
-  render() {
-    const { prefix, id1 } = this.state;
-    return (
-      <Fragment>
-        <EuiFlexGroup
-          justifyContent="flexStart"
-          gutterSize="m"
-          alignItems="center">
-          <EuiFlexItem grow={false}>
-            <EuiFormRow label="Prefix">
-              <EuiFieldText
-                value={prefix}
-                onChange={this.onSearchChange}
-                placeholder="Enter prefix"
-              />
-            </EuiFormRow>
-          </EuiFlexItem>
-        </EuiFlexGroup>
-        <EuiSpacer size="xl" />
-        <EuiCode>{id1} </EuiCode>
-      </Fragment>
-    );
-  }
-}
+  return (
+    <Fragment>
+      <EuiFlexGroup
+        justifyContent="flexStart"
+        gutterSize="m"
+        alignItems="center">
+        <EuiFlexItem grow={false}>
+          <EuiFormRow label="Prefix">
+            <EuiFieldText
+              value={prefix}
+              onChange={onSearchChange}
+              placeholder="Enter prefix"
+            />
+          </EuiFormRow>
+        </EuiFlexItem>
+      </EuiFlexGroup>
+      <EuiSpacer size="xl" />
+      <EuiCode>{customId} </EuiCode>
+    </Fragment>
+  );
+};