+++ src/components/tabs/tabbed_content/tabbed_content.tsx
@@ -66,7 +66,7 @@ export class EuiTabbedContent extends Component<
 
   private readonly rootId = htmlIdGenerator()();
 
-  private readonly divRef = createRef<HTMLDivElement>();
+  private readonly tabsRef = createRef<HTMLDivElement>();
 
   constructor(props: EuiTabbedContentProps) {
     super(props);
@@ -89,9 +89,9 @@ export class EuiTabbedContent extends Component<
   componentDidMount() {
     // IE11 doesn't support the `relatedTarget` event property for blur events
     // but does add it for focusout. React doesn't support `onFocusOut` so here we are.
-    if (this.divRef.current) {
+    if (this.tabsRef.current) {
       // Current short-term solution for event listener (see https://github.com/elastic/eui/pull/2717)
-      this.divRef.current.addEventListener(
+      this.tabsRef.current.addEventListener(
         'focusout' as 'blur',
         this.removeFocus
       );
@@ -99,9 +99,9 @@ export class EuiTabbedContent extends Component<
   }
 
   componentWillUnmount() {
-    if (this.divRef.current) {
+    if (this.tabsRef.current) {
       // Current short-term solution for event listener (see https://github.com/elastic/eui/pull/2717)
-      this.divRef.current.removeEventListener(
+      this.tabsRef.current.removeEventListener(
         'focusout' as 'blur',
         this.removeFocus
       );
@@ -113,7 +113,7 @@ export class EuiTabbedContent extends Component<
       // Must wait for setState to finish before calling `.focus()`
       // as the focus call triggers a blur on the first tab
       this.setState({ inFocus: true }, () => {
-        const targetTab: HTMLDivElement | null = this.divRef.current!.querySelector(
+        const targetTab: HTMLDivElement | null = this.tabsRef.current!.querySelector(
           `#${this.state.selectedTabId}`
         );
         targetTab!.focus();
@@ -170,12 +170,13 @@ export class EuiTabbedContent extends Component<
     const { content: selectedTabContent, id: selectedTabId } = selectedTab!;
 
     return (
-      <div
-        ref={this.divRef}
-        className={className}
-        {...rest}
-        onFocus={this.initializeFocus}>
-        <EuiTabs expand={expand} display={display} size={size}>
+      <div className={className} {...rest}>
+        <EuiTabs
+          ref={this.tabsRef}
+          expand={expand}
+          display={display}
+          size={size}
+          onFocus={this.initializeFocus}>
           {tabs.map((tab: EuiTabbedContentTab) => {
             const {
               id,